Fixed windows build warning

+++ src/runtime_src/xdp/profile/database/static_info/filetypes/aie_trace_config_v3_filetype.cpp
@@ -137,7 +137,7 @@ AIETraceConfigV3Filetype::getTiles(const std::string& graph_name,
 
         // Get core tile location
         auto coreCol = mapping.second.get<uint8_t>("column");
-        auto coreRow = mapping.second.get<uint8_t>("row") + rowOffset;
+        auto coreRow = static_cast<uint8_t>(mapping.second.get<uint8_t>("row") + rowOffset);
 
         // Create or get existing core tile
         auto coreKey = std::make_pair(coreCol, coreRow);
@@ -155,7 +155,7 @@ AIETraceConfigV3Filetype::getTiles(const std::string& graph_name,
         if (dmaChannelsTree) {
             for (auto const &channel : dmaChannelsTree.get()) {
                 uint8_t dmaCol = xdp::aie::convertStringToUint8(channel.second.get<std::string>("column"));
-                uint8_t dmaRow = xdp::aie::convertStringToUint8(channel.second.get<std::string>("row")) + rowOffset;
+                uint8_t dmaRow = static_cast<uint8_t>(xdp::aie::convertStringToUint8(channel.second.get<std::string>("row")) + rowOffset);
 
                 // If channel matches core tile location, add to core tile
                 if (dmaCol == coreCol && dmaRow == coreRow) {
